pyproject.toml
[tool.poetry]
name = "test-2-marusich"
version = "0.1.0"
description = ""
authors = ["polenom <vitalimitsevich@yandex.by>"]
readme = "README.md"
packages = [{include = "test_2_marusich"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
coverage = "^6.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
